backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
django-filter==23.5
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
redis==5.0.1
asgiref>=3.8.1

